Feat: Use cloneElement to get component info in BoardItem
